# api.yaml
swagger: '2.0'
info:
  title: Reshigo API
  version: "1.0"
# the domain of the service
host: server.reshigo.ru
...
schemes:
  - http
# will be prefixed to all paths
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: Http Basic Authentication. Works over `HTTP`
paths:
...
  /user:
    get:
      description: Get user info
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
    put:
      description: Update user\'s info and education and degree fields
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: user
          in: body
          description: Only some of info, education, degree, iosReview, androidReview fields required. Max length is 1000 bytes for each string field
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: OK
        400:
          description: Wrong data in request.
...
  /user/{name}:
    get:
      description: |
        Get user\'s information. Only name, info, education and rating presented.
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: user not found
...
  /user/orders:
    get:
      description: Array of orders associated with the user
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: default offset is 0
          type: integer
          format: int64
        - name: limit
          in: query
          required: false
          description: default limit is 30
          type: integer
          format: int64
        - name: status
          in: query
          required: false
          description: Filter orders by status. Default value is AVAILABLE. |
            Also you can use keywords ACTIVE and NON_ACTIVE and separate statuses by comma like AVAILABLE,SOLVED,DONE
          type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
    post:
      description: Make order available for solvers. Remember, that min price is 80 RUB.
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: order
          description: place here metadata about the order besides pictures
          in: body
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/Order'
        400:
          description: The order is already available, reserved or solved.
        403:
          description: Not enough available funds.
        404:
          description: Check the id of the order. Maybe you do not own this order.
  /user/orders/{id}:
    get:
      description: Information about a concrete order
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Order'
        404:
          description: Not found
    delete:
      description: Remove order. Available only if the order hasn't been reserved by solver yet.
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Removed
        403:
          description: |
            You are not allowed to remove this order.
            The order doesn't exist or has already been reserved.
        404:
          description: The order was not found among your orders.
  /user/orders/{id}/reject:
    post:
      description: Reject solver\'s solution
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        202:
          description: OK
        400:
          description: This action is not available for the order in current state.
        404:
          description: The order was not found
  /user/orders/{id}/done:
    post:
      description: Accept solver\'s solution.
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        202:
          description: OK
        400:
          description: This action is not available for the order in current state.
        404:
          description: The order was not found
  /user/orders/{id}/solve:
    post:
      description: Mark order solved
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        202:
          description: OK
        400:
          description: This action is not available for the order in current state.
        404:
          description: The order was not found
  /user/orders/{id}/appeal:
    post:
      description: Appeal rejected order
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        202:
          description: OK
        400:
          description: This action is not available for the order in current state.
        404:
          description: The order was not found
  /user/orders/{id}/confirm:
    post:
      description: Confirm that order was rejected correctly
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        202:
          description: OK
        400:
          description: This action is not available for the order in current state.
        404:
          description: The order was not found
  /user/orders/{id}/suggest:
    put:
      description: Confirm price suggestion
      tags:
        - Consumer
      security:
      - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
        - name: suggest_id
          in: query
          required: true
          description: suggest id, that user received
          type: integer
      responses:
        202:
          description: OK
        400:
          description: Some error with params. Maybe wrong feed id.
        404:
          description: The order was not found
        405:
          description: Solver is busy at the moment.
        406:
          description: Not enough funds to confirm.
  /user/orders/{id}/suggestions:
    get:
      description: Get price suggestions for the order
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: order's id
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/PriceSuggestion'
  /user/orders/{id}/pictures:
    post:
      description: Add picture for an order that was not sent to moderation or you expired limit of 3 pictures per order.
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
        - name: picture
          in: body
          required: true
          schema:
            $ref: '#/definitions/Picture'
      responses:
        201:
          description: Created. \'img\' property of the return is emply
          schema:
            $ref: '#/definitions/Picture'
        400:
          description: Order was already sent to moderation.
        404:
          description: Order was not found
  /user/orders/{id}/pictures/v2:
      post:
        description: |
          | Deprecated. Use /orders/{id}/pictures/{picture_id}/v2
            | Advanced method to add pictures. Should be much faster. Use "Content-type: multipart/form-data" and put binary data of image in field "file"
        tags:
          - Consumer
        security:
          - basicAuth: []
        consumes: [ "multipart/form-data" ]
        parameters:
          - name: id
            in: path
            required: true
            type: integer
            format: int64
          - name: counter
            in: query
            required: true
            type: integer
            format: int64
        responses:
          201:
            description: Created. \'img\' property of the return is emply
            schema:
              $ref: '#/definitions/Picture'
          400:
            description: Order was already sent to moderation.
          404:
            description: Order was not found
  /user/orders/{id}/maturity:
    put:
      description: Set new maturity date for the order. New maturity = old one + diff (diff > 0).
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          in: path
          required: true
        - name: diff
          type: integer
          format: int64
          in: query
          required: true
      responses:
        200:
          description: OK
        400:
          description: New date is less than existing.
        404:
          description: The order was not found.
  /user/orders/{id}/pictures/{picture_id}:
    get:
      description: Deprecated. Use /orders/{id}/pictures/{picture_id}. Get picture of the order with identifyer 'id'
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
        - name: picture_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Picture'
        404:
          description: Order or picture with given id not found
    delete:
      description: Remove picture from order
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
        - name: picture_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Removed
        400:
          description: The order was not found among your orders.
        404:
          description: The picture was not found.
        405:
          description: The order is not in DRAFTs.
...
  /orders:
    get:
      description: Get available for solving orders.
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: offsetDate
          in: query
          required: false
          description: default value if 2000000000000 (about 2033 year). There are orders with it\'s date less or equal than specified.
          type: integer
          format: int64
        - name: offset
          in: query
          required: false
          description: default offset is 0
          type: integer
          format: int64
        - name: limit
          in: query
          required: false
          description: default limit is 30
          type: integer
          format: int64
        - name: subject
          in: query
          description: |
            subject\'s id, no filter will be applied if not set
          type: integer
          format: int
          required: false
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
  /orders/{id}:
    get:
      description: Get single order from AVAILABLE
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          required: true
          in: path
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Order'
        404:
          description: No such order among AVAILABLE
    put:
      description: Reserve order to solve later. Limit is 2 orders.
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          required: true
          in: path
      responses:
        200:
          description: OK
        403:
          description: The order has already been reserved.
        405:
          description: Solver can't reserve more orders, or this order is only for price suggestions
  /orders/{id}/suggest:
    post:
      summary: Suggest price for order.
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          required: true
          in: path
        - name: price
          in: query
          required: true
          type: integer
        - name: comment
          in: body
          format: string
      responses:
        200:
          description: OK
        403:
          description: Order is not AVAILABLE any more
        404:
          description: Order was not found
        405:
          description: Price is less than 80
  /orders/{id}/pictures:
    get:
      description: Get order\'s pictures, without binary data (img field)
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          in: path
          required: true
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Picture'
          description: OK
        403:
          description: the order is not available
        404:
          description: the order was not found
  /orders/{id}/pictures/{picture_id}:
    get:
          description: Get order\'s picture
          tags:
            - Solver
          security:
            - basicAuth: []
          parameters:
            - name: id
              type: integer
              format: int64
              in: path
              required: true
            - name: picture_id
              type: integer
              format: int64
              in: path
              required: true
          responses:
            200:
              schema:
                $ref: '#/definitions/Picture'
              description: OK
            404:
              description: the order or the picture was not found
  /orders/{id}/pictures/{picture_id}/v2:
      get:
          description: Get order\'s picture as byte stream
          produces:
            - image/jpeg
          tags:
            - Solver
          security:
            - basicAuth: []
          parameters:
            - name: id
              type: integer
              format: int64
              in: path
              required: true
            - name: picture_id
              type: integer
              format: int64
              in: path
              required: true
          responses:
            200:
              description: OK
            404:
              description: the order or the picture was not found
...
definitions:
...
  User:
    type: object
    properties:
      name:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+ and can't consist only from numbers
      email:
        type: string
        description: DEPRECATED
      phone:
        type: string
      fcmToken:
        description: token to use for notifications server
        type: string
      isCustomer:
        type: boolean
      promocode:
        description: sequence of characters to invite friends
        type: string
      funds:
        type: number
        format: double
      reservedFunds:
        type: number
        format: double
      commission:
        type: number
        format: float
      info:
        type: string
        format: binary
        description: utf-8 string
      education:
        type: string
        description: utf-8 string
      degree:
        type: string
        description: utf-8 string, representing class, course or other academic degree
      campaign:
        type: string
        description: source, from where this user came
      iosReview:
        type: string
        description: if true, then user reviewed application in AppStore. can be changed via PUT /user method
      androidReview:
        type: string
        description: if true, then user reviewed application in PlayStore. can be changed via PUT /user method
...
  Picture:
    type: object
    properties:
      id:
        type: integer
        format: int64
      counter:
        type: integer
        format: int64
        description: Field to sort pictures. You had better use unix-timestamp to set this filed.
      img:
        $ref: '#/definitions/PictureData'
  PictureData:
    type: object
    properties:
      data:
        type: string
        format: binary
        description: base64 binary representation of the image
  Order:
    type: object
    description: Every property of theme except \'id\' is ignored.
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
      date:
        type: integer
        format: int64
        description: unix time
      maturityDate:
        type: integer
        format: int64
        description: unix time
      comment:
        type: string
        format: binary
        description: base64 binary representation of string
      theme:
        $ref: '#/definitions/Theme'
      pictures:
        description: This field contains only ids and counters of pictures.
        type: array
        items:
          $ref: '#/definitions/Picture'
      price:
        description: min value is 90
        type: integer
        format: int64
      tasksCnt:
        description: number of tasks
        type: integer
        format: int64
      chat:
        $ref: '#/definitions/Chat'
      user:
        description: Contains customer\'s login and flag isCustomer
        $ref: '#/definitions/User'
      solver:
        description: Contains solver\'s login and flag isCustomer
        $ref: '#/definitions/User'
  Subject:
    type: object
    properties:
      id:
        type: integer
        format: int64
      subject:
        type: string
      subjectRU:
        type: string
        format: binary
  Theme:
    type: object
    properties:
      id:
        type: integer
        format: int64
      theme:
        type: string
      themeRU:
        type: string
        format: binary
      subject:
        $ref: '#/definitions/Subject'
  Chat:
    type: object
    description: All properties of order except id are ignored
    properties:
      id:
        type: integer
        format: int64
      order:
        $ref: '#/definitions/Order'
      status:
        type: string
        description: OPEN, CLOSED
...
  PriceSuggestion:
    type: object
    description: price suggest from solver
    properties:
      id:
        type: integer
        format: int64
      price:
        type: integer
        format: int64
      comment:
        type: string
      order:
        $ref: '#/definitions/Order'
        description: contains only id
      solver:
        $ref: '#/definitions/User'
...